pyproject.toml
[tool.poetry]
name = "papermerge-core"
version = "3.3"
description = "Document management system for digital archives"
authors = ["Eugen Ciur <eugen@papermerge.com>"]
maintainers = ["Eugen Ciur <eugen@papermerge.com>"]
license = "Apache 2.0"
readme = "README.md"
homepage = "https://www.papermerge.com/"
documentation = "https://docs.papermerge.io/"
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3"
]
keywords = [
  "PDF",
  "OCR",
  "optical character recognition",
  "PDF/A",
  "scanning",
  "dms",
  "document-management",
  "paperless",
  "archives"
]
packages = [
    { include = "papermerge/core" },
    { include = "papermerge/search" },
]

[project.urls]
Tracker = "https://github.com/ciur/papermerge/issues"
Source = "https://github.com/papermerge/papermerge-core"

[tool.poetry.dependencies]
python = ">=3.13, <4.0"
redis = "^5.2"
celery = {version = "^5.2", extras = ["gevent", "redis"]}
pyyaml = "^6.0"
python-magic = "^0.4"
uritemplate = "^4.1.1"
pdf2image = "^1.16.0"
uvicorn = "^0.32"
python-jose = "^3.3.0"
salinic = "^0.6.0"
img2pdf = "^0.5.1"
passlib = "^1.7.4"
sqlalchemy = "^2.0"
typer = "^0.13"
rich = "^13"
pytz = "^2024.1"
pydantic-settings = "^2"
alembic = "^1.13.3"
fastapi = {extras = ["standard"], version = "^0.115"}
psycopg2 = { version = "^2.9.2", optional = true}
botocore = { version = "^1.35", optional = true }
boto3 = { version = "^1.35", optional = true }
mysqlclient = {version = "^2.2", optional = true}
cryptography = { version = "^43", optional = true }
taskipy = "^1.14"
pyjwt = "^2.10.1"
django = "^5.1.4"
python-docx = "^1.1.2"
pandas = "^2.2.3"
fpdf = "^1.7.2"
openpyxl = "^3.1.5"
pypandoc = "^1.14"
reportlab = "^4.2.5"
pillow = "^11.0.0"

[tool.poetry.extras]
mysql = ["mysqlclient"]
pg = ["psycopg2"]
databases = ["mysqlclient", "psycopg2"]
cloud = ["botocore", "boto3", "cryptography"]

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^7.1.2"
pytest-asyncio = "^0.21.0"

[tool.taskipy.settings]
use_vars = true

[tool.taskipy.variables]
env_vars = "PAPERMERGE__MAIN__API_PREFIX=''"

[tool.taskipy.tasks]
test = {cmd = "{env_vars} pytest", cwd = "." }
# run following commands from docker/dev only
worker = "celery -A config worker -c 5"
server = "fastapi dev papermerge/app.py"
migrate = "alembic upgrade head"

[tool.poetry.scripts]
paper-cli = "papermerge.cli:app"
docs = 'papermerge.core.cli.docs:app'
sign_url = 'papermerge.core.cli.cf_sign_url:app'


[tool.ruff]
target-version = "py311"
# say hello to balck rigidity
line-length = 88  # same as black's hardcoded
indent-width = 4

[tool.ruff.lint]
"select" = [
    "F",    # pyflakes
    "I001", # isort
    "UP",   # pyupgrade
]
"ignore" = [
    "UP007"
]


[build-system]
requires = ["setuptools >= 61.0", "wheel"]
build-backend = "setuptools.build_meta"
